analyzer/requirements.txt
redis
pymongo
fastapi
uvicorn[standard]
pandas
requests
python-dotenv
